pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastband-mcp"
version = "1.2025.12.0"
description = "Universal AI-powered development platform - AI-agnostic MCP server with adaptive tools, ticket management, and multi-agent coordination"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Fastband Team" }
]
keywords = [
    "mcp",
    "ai",
    "development",
    "tools",
    "automation",
    "tickets",
    "claude",
    "openai",
    "gemini",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "mcp>=0.9.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
claude = [
    "anthropic>=0.18.0",
]
openai = [
    "openai>=1.0.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]
ollama = [
    "ollama>=0.1.0",
]
all-providers = [
    "fastband-mcp[claude,openai,gemini,ollama]",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "jinja2>=3.0.0",
    "websockets>=11.0.0",
]
hub = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0.0",
    "aiofiles>=23.0.0",
    "numpy>=1.24.0",
]
screenshots = [
    "playwright>=1.40.0",
    "pillow>=10.0.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
mysql = [
    "pymysql>=1.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "numpy>=1.24.0",
]
full = [
    "fastband-mcp[all-providers,web,hub,screenshots,postgres,mysql,dev]",
]

[project.scripts]
fastband = "fastband.__main__:main"
fb = "fastband.__main__:main"

[project.urls]
Homepage = "https://github.com/RemmyCH3CK/fastband-mcp"
Documentation = "https://github.com/RemmyCH3CK/fastband-mcp#readme"
Repository = "https://github.com/RemmyCH3CK/fastband-mcp"
Issues = "https://github.com/RemmyCH3CK/fastband-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/fastband"]

# Include built React dashboard static files in the package
# The hub/static/ directory is created by `fastband build-dashboard`
[tool.hatch.build.targets.wheel.force-include]
"src/fastband/hub/static" = "fastband/hub/static"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "B904",   # raise-without-from-inside-except (style preference)
    "E741",   # ambiguous-variable-name (using l, O, I)
    "F841",   # unused variable (addressed separately)
    "F401",   # unused-import (re-exports in __init__.py)
    "UP035",  # deprecated-import (typing.Dict -> dict)
    "E402",   # module-import-not-at-top (sometimes needed)
    "B017",   # assert-raises-exception (testing patterns)
    "B027",   # empty-method-without-abstract-decorator
    "B007",   # unused-loop-control-variable
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=fastband --cov-report=term-missing"

[tool.coverage.run]
source = ["src/fastband"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
